Update main.bat
in this update i added a function that search through the files to find AngryIp.exe. Of course the data on your computer stay on your computer and isnt going to be sent to me. Once the program is on your computer the program is local. That doesnt mean you are allowed to steal, copy, modify my code without giving me credits. If you dont follow the license then ill have no choice but to defend myself